Add deserialization
Add deserialization and test on availability for PUT and DELETE methods

+++ app.py
@@ -1,30 +1,21 @@
-from flask import Flask, request, jsonify
-#from service import ToDoService
-#from models import Schema
+from flask import Flask, request, jsonify, abort
 from flask_sqlalchemy import SQLAlchemy
 from flask_marshmallow import Marshmallow
-import os
 import json
 
-
-
 app = Flask(__name__)
-basedir = os.path.abspath(os.path.dirname(__file__))
-#app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///' + os.path.join(basedir, 'goods.sqlite')
-#app.config['SQLALCHEMY_DATABASE_URI'] = 'mysql://alexf:Alex123456788!@localhost/lab_6_db'format(username, password, server)
-DATABSE_URI='mysql+mysqlconnector://alexf:Alex123456788!@localhost/lab_6_db'.format(user='alexf', password='Alex123456788!', server='localhost', database='lab_6_db')
+DATABSE_URI='mysql+mysqlconnector://alexf:Alex123456788!@localhost/lab_6_db'
 
 db = SQLAlchemy(app)
-
 ma = Marshmallow(app)
 
 class Good(db.Model):
     id = db.Column(db.Integer, primary_key=True)
-    name = db.Column(db.String(80), unique=True)
-    price = db.Column(db.Float, unique=True)
-    weight = db.Column(db.Float, unique=True)
-    producer = db.Column(db.String(80), unique=True)
-    country = db.Column(db.String(80), unique=True)
+    name = db.Column(db.String, unique=True)
+    price = db.Column(db.Float)
+    weight = db.Column(db.Float)
+    producer = db.Column(db.String)
+    country = db.Column(db.String)
 
 
     def __init__(self, name, price, weight, producer, country):
@@ -35,7 +26,7 @@ class Good(db.Model):
         self.country = country
 
     def __repr__(self):
-        return '<User %r>' % self.name
+        return '<Good %r>' % self.name
 
 class GoodSchema(ma.Schema):
     class Meta:
@@ -50,24 +41,17 @@ db.create_all()
 # endpoint to create new good
 @app.route("/good", methods=["POST"])
 def add_good():
-    name = request.json['name']
-    price = request.json['price']
-    weight = request.json['weight']
-    producer = request.json['producer']
-    country = request.json['country']
-    
-    new_good = Good(name, price, weight, producer, country)
+    data = GoodSchema().load(request.json)
+    new_good = Good(**data)
 
     db.session.add(new_good)
     db.session.commit()
-
-    return jsonify(new_good)
+    return "Added : " + new_good.name
 
 # endpoint to show all goods
 @app.route("/all_goods", methods=["GET"])
 def get_goods():
     goods = Good.query.all()
-    goods_schema = GoodSchema(many=True)
     output = goods_schema.dump(goods)
     return jsonify({'goods': output})
 
@@ -75,29 +59,25 @@ def get_goods():
 @app.route("/good/<id>", methods=["PUT"])
 def good_update(id):
     good = Good.query.get(id)
-    name = request.json['name']
-    price = request.json['price']
-    weight = request.json['weight']
-    producer = request.json['producer']
-    country = request.json['country']
-
-    good.name = name
-    good.price = price
-    good.weight = weight
-    good.producer = producer
-    good.country = country
-
-    db.session.commit()
-    return good_schema.jsonify(good)
+    if good is None:
+        abort(404)
+    else:
+        data = good_schema.load(request.json)
+        for i in data:
+            setattr(good, i, request.json[i])
+        db.session.commit()
+        return good_schema.jsonify(good)
 
 # endpoint to delete good
 @app.route("/good/<id>", methods=["DELETE"])
 def good_delete(id):
     good = Good.query.get(id)
-    db.session.delete(good)
-    db.session.commit()
-
-    return good_schema.jsonify(good)
+    if good is None:
+        abort(404)
+    else:
+        db.session.delete(good)
+        db.session.commit()
+        return "Deleted: " + good.name
 
 
 if __name__ == '__main__':
